refactor(engine-cli): remove direct require usages preparations for more esm
